render/allocator: make wlr_allocator part of the public API
