+++ render/allocator/shm.c
@@ -3,7 +3,10 @@
 #include <stdlib.h>
 #include <sys/mman.h>
 #include <unistd.h>
+#include <wlr/render/allocator.h>
+#include <wlr/render/drm_format_set.h>
 #include <wlr/util/log.h>
+
 #include "render/pixel_format.h"
 #include "render/allocator/shm.h"
 #include "util/shm.h"
